+++ index.js
@@ -1,0 +1,4 @@
+require('./launchServices');
+require('./webStreams');
+require('./appCtrl');
+require('./server').listen(15000);
