+++ server.js
@@ -1,0 +1,3 @@
+var http = require('http');
+
+module.exports = http.createServer();
